in progress working version from scratch
