+++ src/app/app.module.ts
@@ -6,7 +6,6 @@ import { ToastrModule } from 'ngx-toastr';
 import { AppComponent } from './app.component';
 import { CounterService } from './services/counter.service';
 import { HttpModule } from '@angular/http';
-import { BodyComponent } from './body/body.component';
 import { FooterComponent } from './footer/footer.component';
 import { AddCountryComponent } from './components/add-country/add-country.component';
 import { CountrySelectorComponent } from './components/country-selector/country-selector.component';
@@ -14,7 +13,6 @@ import { CountrySelectorComponent } from './components/country-selector/country-
 @NgModule({
   declarations: [
     AppComponent,
-    BodyComponent,
     FooterComponent,
     AddCountryComponent,
     CountrySelectorComponent
